[Refactor] test structure
