+++ paper_cup/test/test_paper_cup.py
@@ -5,25 +5,6 @@ from paper_cup.client import SNSClient, SQSClient
 from paper_cup.paper_cup import PaperCup
 
 
-class TestSNSClient(TestCase):
-
-  def test_instance(self):
-    """Check that the instance initialize correctly."""
-    sns_client = SNSClient()
-    # check that we correctly set an sns session
-    self.assertTrue(sns_client.sns)
-
-
-class TestSQSClient(TestCase):
-
-  def test_instance(self):
-    """Check that the instance initialize correctly."""
-    sqs_client = SQSClient()
-    # check that we correctly set an sqs session
-    self.assertTrue(sqs_client.sqs)
-    self.assertTrue(sqs_client.sqs_obj)
-
-
 class TestPaperCup(TestCase):
 
   def setUp(self):
